Move BaseMetadataMutation to deprecated folder
